+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/graph/BindingGraph.kt
@@ -61,7 +61,11 @@ internal open class MutableBindingGraph<
     },
   private val onError: (String, BindingStack) -> Unit = { message, _ -> error(message) },
   private val onHardError: (String, BindingStack) -> Nothing = { message, _ -> error(message) },
-  private val findSimilarBindings: (key: TypeKey) -> Map<TypeKey, String> = { emptyMap() },
+  private val missingBindingHints:
+    (key: TypeKey, stack: BindingStack) -> MissingBindingHints<Type, TypeKey> =
+    { _, _ ->
+      MissingBindingHints()
+    },
 ) : BindingGraph<Type, TypeKey, ContextualTypeKey, Binding, BindingStackEntry, BindingStack> {
   // Populated by initial graph setup and later seal()
   override val bindings = mutableMapOf<TypeKey, Binding>()
@@ -442,6 +446,7 @@ internal open class MutableBindingGraph<
     bindingStack: BindingStack,
     extraContent: StringBuilder.() -> Unit = {},
   ) {
+    // TODO referents
     if (reportedMissingKeys.add(typeKey)) {
       val message = buildString {
         append(
@@ -450,12 +455,23 @@ internal open class MutableBindingGraph<
         appendLine(typeKey.render(short = false))
         appendLine()
         appendBindingStack(bindingStack, short = false)
-        val similarBindings = findSimilarBindings(typeKey)
-        if (similarBindings.isNotEmpty()) {
-          appendLine()
-          appendLine("Similar bindings:")
-          similarBindings.values.map { "  - $it" }.sorted().forEach(::appendLine)
+        val hints = missingBindingHints(typeKey, bindingStack)
+        val messages = hints.messages
+        val similarBindings = hints.similarBindings
+
+        if (messages.isNotEmpty() || similarBindings.isNotEmpty()) {
+          if (messages.isNotEmpty()) {
+            appendLine()
+            appendLine("(Hint)")
+            messages.joinTo(this, separator = "\n\n")
+          }
+          if (similarBindings.isNotEmpty()) {
+            appendLine()
+            appendLine("Similar bindings:")
+            similarBindings.values.map { "  - $it" }.sorted().forEach(::appendLine)
+          }
         }
+
         extraContent()
       }