+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/util.kt
@@ -9,11 +9,13 @@ import org.jetbrains.kotlin.name.ClassId
 import org.jetbrains.kotlin.name.Name
 
 private val PLATFORM_TYPE_PACKAGES =
-  setOf("android.", "androidx.", "java.", "javax.", "kotlin.", "kotlinx.", "scala.")
+  setOf("android", "androidx", "java", "javax", "kotlin", "kotlinx", "scala")
 
 internal fun ClassId.isPlatformType(): Boolean {
   return packageFqName.asString().let { packageName ->
-    PLATFORM_TYPE_PACKAGES.any { packageName.startsWith(it) }
+    PLATFORM_TYPE_PACKAGES.any { platformPackage ->
+      packageName == platformPackage || packageName.startsWith("$platformPackage.")
+    }
   }
 }